requirements.txt
fastapi<0.110
pydantic<2.0.0
uvicorn
sqlalchemy
python-dotenv
psycopg2-binary
